implemented rename_path and tests (sync)

+++ sdk/storage/azure-storage-file-datalake/tests/test_directory.py
@@ -1637,22 +1637,33 @@ class TestDirectory(StorageRecordedTestCase):
         datalake_storage_account_name = kwargs.pop("datalake_storage_account_name")
         datalake_storage_account_key = kwargs.pop("datalake_storage_account_key")
 
-        # Arrange
         self._setUp(datalake_storage_account_name, datalake_storage_account_key)
         directory_name = self._get_directory_reference()
         directory_client = self.dsc.get_directory_client(self.file_system_name, directory_name)
+        new_directory_client = None
         create_transaction_id = 'd9a752bb-3702-4336-b4fb-470bdfad'
+        rename_transaction_id = '6a9a3e11-5074-42f3-a15e-d445548d'
 
         try:
-            # Act
             directory_client.create_directory(client_transaction_id=create_transaction_id)
-
-            # Assert
             props = directory_client._client.path.get_properties(cls=return_response_headers)
             assert props is not None
             assert props["client_transaction_id"] == create_transaction_id
+
+            new_directory_client = directory_client.rename_directory(
+                self.file_system_name + '/' + 'newdir',
+                client_transaction_id=rename_transaction_id
+            )
+            props = new_directory_client._client.path.get_properties(cls=return_response_headers)
+            assert props is not None
+            assert props["client_transaction_id"] == rename_transaction_id
+
+            directory_client = None
         finally:
-            directory_client.delete_directory()
+            if directory_client is not None:
+                directory_client.delete_directory()
+            if new_directory_client is not None:
+                new_directory_client.delete_directory()
 
 # ------------------------------------------------------------------------------
 if __name__ == '__main__':
